@@ -7,7 +7,7 @@ publish_to: none
 
 
 environment:
-  sdk: '>=2.10.5 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   collection: ^1.15.0-nullsafety.4
